２要素とログイン履歴ページをまとめる
